# docs/api/openapi.yaml
openapi: 3.0.0
paths:
...
  /v1/plans:
    get:
      operationId: PlansController_list
      parameters: []
      responses:
        "200":
          description: ""
      tags: &a2
        - Plans
    post:
      operationId: PlansController_create
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePlanDto"
      responses:
        "201":
          description: ""
      tags: *a2
...
  /v1/plans/{planId}/sections:
    get:
      operationId: PlansController_listSections
      parameters:
        - name: planId
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: ""
      tags: *a2
    post:
      operationId: PlansController_createSection
      parameters:
        - name: planId
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSectionDto"
      responses:
        "201":
          description: ""
      tags: *a2
...
info:
  title: IREAL Service API
  description: Versioned API for the decoupled IREAL backend service.
  version: 0.0.1
  contact: {}
...
servers: []
components:
  schemas:
...
    CreatePlanDto:
      type: object
      properties: {}
...
    CreateSectionDto:
      type: object
      properties: {}
...